Update the README.
It now mentions 'dynamic patching' and reads a bit better in some areas.
Unfortunately we've grown by ~60 kB (not that this is a big problem). Oh
also there's a disclaimer about packs actually needing to support
Patched for there to be any differences, since otherwise it's easy to
read that section as suggesting Patched will magically make your data
packs compatible with each other.