Migrate documentation to a GitHub wiki.
There are two reasons for this: first, the wiki is easier to navigate,
since the sidebar is always visible; second, it removes the need to
maintain multiple versions of the documentation.

It's still a little rough around the edges, although this argument can
also be applied to the docs as they were before. It would not be a
terrible idea to rewrite some of them for improved organization and
clarity. (And bring the library's docs over so people can have more
relevant examples!)