[Breaking Change] Fix patch application order.
Previously, patches were applied in order of highest to lowest priority.
This works fine for regular files but no so much for patches. Now
patches are applied in order of lowest to highest priority, allowing one
to write a patch that 'patches out' the changes made by a lower-priority
patch.